consolidate tests in single spec file
